requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.2
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
iperf3==0.1.11
numpy==1.23.2
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
pyzbar==0.1.9
qrcode==7.3.1
requests==2.28.1
selenium==4.4.3
sniffio==1.2.0
sortedcontainers==2.4.0
speedtest-cli==2.1.3
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.8.3
wsproto==1.1.0
XlsxWriter==3.0.3
